Added dec2bin to fix SPI configuration -- still problems
Still problems with the SPI clock phase configuration setting.  I'm
sending the right bits, but I'm not getting a success indication.
Everyone tells me to update the firmware, but it's not clear how to
get the latest version, and what the fixes are.

In the end, the documentation is bad.  Even if I could configure those
bits, sending data over SPI using the bitbang protocol isn't well
documented.